+++ src/config.js
@@ -1,0 +1,1 @@
+local.workerBootstrapUrl = 'worker.min.js';