+++ CHANGES.md
@@ -10,6 +10,9 @@ Changes
    - Changed local.client.listen to local.bindRequestEvents
    - Changed local.client.unlisten to local.unbindRequestEvents
    - Removed regions and response handling utilities (they will be moved to a separate repo)
+ - Removed local.env
+   - Moved Server classes into local.web
+   - Renamed Server.handleHttpRequest to handleWebRequest
 
 
 2013/08/09 pfraze